+++ store-service/src/main/kotlin/store/dto/product/ProductRequest.kt
@@ -3,13 +3,13 @@ package store.dto.product
 data class CreateProductRequest(
     val name: String,
     val description: String,
-    val price: Int,
+    val price: Long,
     val productPictureUrl: String?
 )
 
 data class UpdateProductRequest(
     val name: String,
     val description: String,
-    val price: Int,
+    val price: Long,
     val productPictureUrl: String?
 )
